Fixes #24246 - Remove lookup keys extensions
Hiding the lookup keys in Parameters tab was introduced
so that users would not modify the port, server and policies
keys, which would lead to misconfigureations. But users can
still do that by creating a matcher when editing puppet
classes.

Several new class params were added
and now that we have fetch-remote-resources, I see a value
of being able to change the params directly when editing host.