Fixes #26693 - Switch Result and Severity in Report
Currently the pass/fail for each rule in the report is in the fourth column and the severity is in the left. In our lab, this confused many people who saw the severity red boxes and thought the rule was still failing. The first column showing the result column should clear up this confusion. This commit switches those columns.