Fixes #31381 - Register REX features in to_prepare block
When openscap tried to register REX features, REX was not registered as a plugin
yet and because it was not registered yet, the feature registration got skipped.

This had two causes, features being registered from within a plugin registration
block and foreman_openscap being registered before remote execution as a plugin.
To_prepare block is run after all the plugin-registering initializers so it
addresses both causes.