Switch to gh actions for tests
We need to test with foreman_ansible
to ensure that logic responsible
for config deployment to client
is tested properly. Our Jenkins currently
does not offer any support for testing
multiple plugins together, so a completely
new job would have to be created in JJB.
Swithching to gh actions is easier
and does not require changes
in a different repo.