Fixes #37396 - Fix autocomplete for ArfReports (#567)
* Fixes #37396 - Fix autocomplete for ArfReports

The code expected that a host would have up to one lifecycle
environment, which doesn't hold anymore on recent versions of katello.
This tripped up scoped_search which couldn't find the right association
to search on.

* Fixes #37396 - Fix search on lifecycle_environments

We have to elaborate a bit on the associations to make scoped_search be
able to put the search together through the entire chain of the associations.

(cherry picked from commit 868769ffe24d429e50e918469845b8dbacacc627)